refactor(PaidSession): Session에서 PaidSession분리
