+++ src/test/java/nextstep/session/domain/PaidSessionTest.java
@@ -13,7 +13,7 @@ import static nextstep.users.domain.fixture.NsUserFixture.STUDENT_2;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class SessionTest {
+class PaidSessionTest {
     LocalDate today;
 
     @BeforeEach
@@ -23,10 +23,10 @@ class SessionTest {
     }
 
     @Test
-    @DisplayName("수강신청 / 유료 정원 2명 / 성공")
+    @DisplayName("수강신청 / 유료 정원 2명, 2명 신청 / 성공")
     void 수강신청_유료_성공() {
         // given
-        Session session = Session.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 2, 1000L);
+        PaidSession session = PaidSession.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 2, 1000L);
         session.changeStatus(SessionStatus.RECRUITING);
         STUDENT_1.addPayment(new Payment(1000L, session, STUDENT_1));
         STUDENT_2.addPayment(new Payment(1000L, session, STUDENT_2));
@@ -43,7 +43,7 @@ class SessionTest {
     @DisplayName("수강신청 / 유료 정원 1명, 2명 신청 / IllegalStateException")
     void 수강신청_유료_정원초과_실패() {
         // given
-        Session session = Session.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
+        PaidSession session = PaidSession.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
         session.changeStatus(SessionStatus.RECRUITING);
         STUDENT_1.addPayment(new Payment(1000L, session, STUDENT_1));
         STUDENT_2.addPayment(new Payment(1000L, session, STUDENT_2));
@@ -60,7 +60,7 @@ class SessionTest {
     @DisplayName("수강신청 / 결제금액 같은 유료강의 / 성공")
     void 수강신청_결제금액같은_유료강의_성공() {
         // given
-        Session session = Session.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
+        PaidSession session = PaidSession.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
         session.changeStatus(SessionStatus.RECRUITING);
         STUDENT_1.addPayment(new Payment(1000L, session, STUDENT_1));
 
@@ -72,10 +72,10 @@ class SessionTest {
     }
 
     @Test
-    @DisplayName("수강신청 / 결제금액 다른 유료강의 / 실패")
+    @DisplayName("수강신청 / 결제금액 다른 유료강의 / IllegalArgumentException")
     void 수강신청_결제금액_다른_유료강의_실패() {
         // given
-        Session session = Session.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
+        PaidSession session = PaidSession.create(1, 1L, today, today.plusDays(1), sessionImageFixture, SessionType.PAID, 1, 1000L);
         session.changeStatus(SessionStatus.RECRUITING);
         STUDENT_1.addPayment(new Payment(1200L, session, STUDENT_1));
 
@@ -83,4 +83,5 @@ class SessionTest {
         assertThatThrownBy(() -> session.enroll(STUDENT_1))
                 .isInstanceOf(IllegalArgumentException.class);
     }
+
 }
